requirements.txt
requests
python-dotenv
pydantic
fastapi
uvicorn
rq
redis
# (solo si usás openai oficial)
openai
